\name{xtensor-package}
\alias{xtensor-package}
\alias{xtensor}
\docType{package}
\title{\packageTitle{xtensor}}
\description{\packageDescription{xtensor}}
%\details{
%  The DESCRIPTION file: \packageDESCRIPTION{xtensor}
%  \packageIndices{xtensor}
%}
\author{
  \packageAuthor{xtensor}
  
  Maintainer: \packageMaintainer{xtensor}
}
%\references{
%}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
%\seealso{
%  Optional links to other man pages
%}
%\examples{
%  ## Optional simple examples of the most important functions
%  ## Use \dontrun{} around code to be shown but not executed
%}
